requirements.txt
beautifulsoup4==4.13.3
Requests==2.32.3
pillow~=11.2.1